+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The following components now support dark mode.
 ### New features
 
 - `form-input`: A new `triggerFocus` method is exposed to allow focusing the underlying input.
+- `donut-chart`: A new `colourful` mode uses the brighter set of chart colours. Use with caution.
 
 ### Changes